Bump canboat -> canboat-1.0.0
This commit bumps the current `canboat` binaries
to canboat's official v1.0.0 release. Previously,
the recipe used SRCREV to obtain the code from the
repository. This commit now uses the github tar ball
for compiling.

Changes: Some changes in the main repository needed
to be addressed in the patches. For instance, previous
occurences of `php shebangs` in some files were removed,
and this needs to be provided in the new patches. A line
in the `Makefile` was causing installation error and a patch
for the same is renamed.

Signed-off-by: Shantanoo <shantanoo.desai@gmail.com>